+++ query/query3_test.go
@@ -18,18 +18,12 @@ package query
 
 import (
 	"context"
-	"encoding/binary"
 	"encoding/json"
 	"fmt"
 	"strings"
 	"testing"
 
-	"github.com/dgraph-io/dgo/v2/protos/api"
-	"github.com/dgraph-io/dgraph/testutil"
 	"github.com/stretchr/testify/require"
-	geom "github.com/twpayne/go-geom"
-	"github.com/twpayne/go-geom/encoding/geojson"
-	"github.com/twpayne/go-geom/encoding/wkb"
 	"google.golang.org/grpc/metadata"
 )
 
@@ -2197,111 +2191,4 @@ func TestMultiRegexInFilter2(t *testing.T) {
 		res := processQueryNoErr(t, query)
 		require.JSONEq(t, `{"data": {"q": [{"firstName": "Han", "lastName":"Solo"}]}}`, res)
 	}
-}
-
-func TestGeoDataInvalidString(t *testing.T) {
-	dg, err := testutil.DgraphClientWithGroot(testutil.SockAddr)
-	require.NoError(t, err)
-
-	ctx := context.Background()
-	require.NoError(t, dg.Alter(ctx, &api.Operation{DropAll: true}))
-	require.NoError(t, dg.Alter(ctx, &api.Operation{Schema: `loc: geo .`}))
-
-	n := &api.NQuad{
-		Subject:   "_:test",
-		Predicate: "loc",
-		ObjectValue: &api.Value{
-			Val: &api.Value_StrVal{
-				StrVal: `{"type": "Point", "coordintaes": [1.0, 2.0]}`,
-			},
-		},
-	}
-	_, err = dg.NewTxn().Mutate(ctx, &api.Mutation{
-		CommitNow: true,
-		Set:       []*api.NQuad{n},
-	})
-	require.Contains(t, err.Error(), "geom: unsupported layout NoLayout")
-}
-
-// This test shows that GeoVal API doesn't accept string data. Though, mutation
-// succeeds querying the data returns an error. Ideally, we should not accept
-// invalid data in a mutation though that is left as future work.
-func TestGeoCorruptData(t *testing.T) {
-	dg, err := testutil.DgraphClientWithGroot(testutil.SockAddr)
-	require.NoError(t, err)
-
-	ctx := context.Background()
-	require.NoError(t, dg.Alter(ctx, &api.Operation{DropAll: true}))
-	require.NoError(t, dg.Alter(ctx, &api.Operation{Schema: `loc: geo .`}))
-
-	n := &api.NQuad{
-		Subject:   "_:test",
-		Predicate: "loc",
-		ObjectValue: &api.Value{
-			Val: &api.Value_GeoVal{
-				GeoVal: []byte(`{"type": "Point", "coordinates": [1.0, 2.0]}`),
-			},
-		},
-	}
-	_, err = dg.NewTxn().Mutate(ctx, &api.Mutation{
-		CommitNow: true,
-		Set:       []*api.NQuad{n},
-	})
-	require.NoError(t, err)
-
-	query := `
-{
-  all(func: has(loc)) {
-      uid
-      loc
-  }
-}`
-	_, err = dg.NewReadOnlyTxn().Query(ctx, query)
-	require.Contains(t, err.Error(), "wkb: unknown byte order: 1111011")
-}
-
-// This test shows how we could use the GeoVal API to store geo data.
-// As far as I (Aman) know, this is something that should not be used
-// by a common user unless user knows what she is doing.
-func TestGeoValidWkbData(t *testing.T) {
-	dg, err := testutil.DgraphClientWithGroot(testutil.SockAddr)
-	require.NoError(t, err)
-
-	ctx := context.Background()
-	require.NoError(t, dg.Alter(ctx, &api.Operation{DropAll: true}))
-	require.NoError(t, dg.Alter(ctx, &api.Operation{Schema: `loc: geo .`}))
-	s := `{"type": "Point", "coordinates": [1.0, 2.0]}`
-	var gt geom.T
-	if err := geojson.Unmarshal([]byte(s), &gt); err != nil {
-		panic(err)
-	}
-	data, err := wkb.Marshal(gt, binary.LittleEndian)
-	if err != nil {
-		panic(err)
-	}
-	n := &api.NQuad{
-		Subject:   "_:test",
-		Predicate: "loc",
-		ObjectValue: &api.Value{
-			Val: &api.Value_GeoVal{
-				GeoVal: data,
-			},
-		},
-	}
-
-	_, err = dg.NewTxn().Mutate(ctx, &api.Mutation{
-		CommitNow: true,
-		Set:       []*api.NQuad{n},
-	})
-	require.NoError(t, err)
-	query := `
-{
-  all(func: has(loc)) {
-      uid
-      loc
-  }
-}`
-	resp, err := dg.NewReadOnlyTxn().Query(ctx, query)
-	require.NoError(t, err)
-	require.Contains(t, string(resp.Json), `{"type":"Point","coordinates":[1,2]}`)
 }
